update flags in cmakelists

+++ tests/test_helpers.cpp
@@ -53,7 +53,7 @@ int countNewlines(const std::string& text) {
 char **create_str_arr(std::vector<std::string> strings) {
 	char** strs = (char**)malloc(sizeof(char*) * strings.size());
 
-	for (int i = 0; i < strings.size(); i++)
+	for (unsigned long i = 0; i < strings.size(); i++)
 		strs[i] = ft_strdup(strings[i].c_str());
 
 	return strs;
